Moving the plotting module.
